// apps/web/src/api/generated/schemas/replacementStatusUpdate.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * X-Ear CRM API
 * Auto-generated from Flask backend routes
 * OpenAPI spec version: 1.0.0
 */
import type { ReplacementStatusUpdateStatus } from './replacementStatusUpdateStatus';
import type { ReplacementStatusUpdateNotes } from './replacementStatusUpdateNotes';

export interface ReplacementStatusUpdate {
  status?: ReplacementStatusUpdateStatus;
  notes?: ReplacementStatusUpdateNotes;
}